+++ mullvad-daemon/src/cli.rs
@@ -7,6 +7,7 @@ pub struct Config {
     pub log_level: log::LogLevelFilter,
     pub log_file: Option<PathBuf>,
     pub tunnel_log_file: Option<PathBuf>,
+    pub resource_dir: Option<PathBuf>,
 }
 
 pub fn get_config() -> Config {
@@ -20,11 +21,13 @@ pub fn get_config() -> Config {
     };
     let log_file = matches.value_of_os("log_file").map(PathBuf::from);
     let tunnel_log_file = matches.value_of_os("tunnel_log_file").map(PathBuf::from);
+    let resource_dir = matches.value_of_os("resource_dir").map(PathBuf::from);
 
     Config {
         log_level,
         log_file,
         tunnel_log_file,
+        resource_dir,
     }
 }
 
@@ -43,12 +46,21 @@ fn create_app() -> App<'static, 'static> {
             Arg::with_name("log_file")
                 .long("log")
                 .takes_value(true)
+                .value_name("PATH")
                 .help("Activates file logging to the given path"),
         )
         .arg(
             Arg::with_name("tunnel_log_file")
                 .long("tunnel-log")
                 .takes_value(true)
+                .value_name("PATH")
                 .help("Save log from tunnel implementation process to this file path"),
+        )
+        .arg(
+            Arg::with_name("resource_dir")
+                .long("resource-dir")
+                .takes_value(true)
+                .value_name("DIR")
+                .help("Uses the given directory to read needed resources, such as certificates."),
         )
 }