+++ CHANGELOG.md
@@ -17,6 +17,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 ### Added
 - Uses the https://am.i.mullvad.net/ service to figure out location and public IP of the device.
   The app then shows this information in the unsecured state.
+- Argument to the daemon, `--resource-dir <path>`, that allows customizing where it will look for
+  needed resource files.
 
 ### Fixed
 - Fixed bug where problem report tool would redact some things in the logs which were not IPv6